Revert "6 split app.py into multiple files with blueprints (#7)"
This reverts commit 363e19830f8b5dd3f21f2736d4d577903a4cacfa.
