+++ config.py
@@ -1,2 +1,2 @@
 class Config:
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///wait_lists.db'
+    DATABASE_URI = 'sqlite:///wait_lists.db'
