+++ requirements.txt
@@ -119,8 +119,3 @@ webencodings==0.5.1
 websocket-client==1.3.3
 Werkzeug==2.2.2
 wrapt==1.14.1
-
-qrcode==7.3.1
-SQLAlchemy==1.4.45 
-flask_sqlalchemy==3.0.2 
-greenlet==2.0.1
