+++ templates/wait_list.html
@@ -34,6 +34,6 @@
     </form>
     {% endif %}
     <a href="/manage">Back to Wait List Management</a>
-    <img src="{{ url_for('main_bp.qr', wait_list_id=wait_list_id) }}">
+    <img src="{{ url_for('qr', wait_list_id=wait_list_id) }}">
 </body>
 </html>
