--- waitlist_app/__init__.py
@@ -1,20 +1,0 @@
-from flask import Flask, Blueprint
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-
-def init_app():
-    app = Flask(__name__, instance_relative_config=False)
-    app.config.from_object('config.Config')
-
-    db.init_app(app)
-
-    with app.app_context():
-        from . import routes
-        
-        app.register_blueprint(routes.main_bp)
-
-        # Create Database Models
-        db.create_all() 
-
-        return app
