--- wsgi.py
@@ -1,7 +1,0 @@
-from waitlist_app import init_app
-
-
-application = init_app()
-
-if __name__ == "__main__":
-    application.run()
